+++ packages/web3-core/src/utils.ts
@@ -54,20 +54,21 @@ export const isLegacySendAsyncProvider = <API extends Web3APISpec>(
 export const isSupportedProvider = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
 ): provider is SupportedProviders<API> =>
-	isWeb3Provider(provider) ||
-	isEIP1193Provider(provider) ||
-	isLegacyRequestProvider(provider) ||
-	isLegacySendAsyncProvider(provider) ||
-	isLegacySendProvider(provider);
+	provider &&
+	(isWeb3Provider(provider) ||
+		isEIP1193Provider(provider) ||
+		isLegacyRequestProvider(provider) ||
+		isLegacySendAsyncProvider(provider) ||
+		isLegacySendProvider(provider));
 
 export const isSupportSubscriptions = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
 ): boolean => {
-	if (isWeb3Provider<API>(provider)) {
+	if (provider && 'supportsSubscriptions' in provider) {
 		return provider.supportsSubscriptions();
 	}
 
-	if (typeof provider !== 'string' && 'on' in provider) {
+	if (provider && typeof provider !== 'string' && 'on' in provider) {
 		return true;
 	}