+++ packages/web3/CHANGELOG.md
@@ -123,8 +123,14 @@ Documentation:
 -   Fixed bug #6185, now web3.js compiles on typescript v5 (#6195)
 -   Fixed #6162 @types/ws issue (#6205)
 
-## [Unreleased]
+## [4.0.3]
+
+## Added
+
+-   Web3 constructor accepts `Web3ContextInitOptions<EthExecutionAPI, CustomRegisteredSubscription>` as alternative to the still supported `undefined`, `string`, and `SupportedProviders<EthExecutionAPI>` (#6262).
 
 ### Fixed
 
 -   Fixed bug #6236 by adding personal type in web3.eth (#6245)
+
+## [Unreleased]